chore: refactor file structures
